src/dl_methods/jamie_experiments/requirements_jm.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
adjusttext=0.8=pypi_0
anndata=0.9.1=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
brokenaxes=0.5.0=pypi_0
ca-certificates=2024.11.26=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.0.1=pypi_0
cloudpickle=2.2.0=pypi_0
cmake=3.31.2=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_1
entrypoints=0.4=pyhd8ed1ab_1
et-xmlfile=1.1.0=pypi_0
executing=2.1.0=pyhd8ed1ab_1
filelock=3.12.2=pypi_0
flatbuffers=22.9.24=pypi_0
fonttools=4.41.0=pypi_0
gast=0.4.0=pypi_0
google-auth=2.12.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.49.1=pypi_0
h5py=3.12.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=6.0.0=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.0=pyh41d4057_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.2=pypi_0
joblib=1.3.1=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libclang=14.0.6=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
lit=18.1.8=pypi_0
llvmlite=0.40.1=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mmd-wrapper=1.2.5=pypi_0
mpmath=1.3.0=pypi_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.1=pypi_0
numba=0.57.1=pypi_0
numpy=1.21.6=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.1=pypi_0
openpyxl=3.0.10=pypi_0
openssl=3.0.15=h5eee18b_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
patsy=0.5.3=pypi_0
pep517=0.11.0=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=10.0.0=pypi_0
pip=24.2=py39h06a4308_0
pip-tools=6.3.0=pypi_0
platformdirs=4.3.6=pyhd8ed1ab_1
prompt-toolkit=3.0.48=pyha770c72_1
prompt_toolkit=3.0.48=hd8ed1ab_1
protobuf=3.19.6=pypi_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.18.0=pyhd8ed1ab_1
pynndescent=0.5.8=pypi_0
pyparsing=3.0.9=pypi_0
python=3.9.21=he870216_1
python-dateutil=2.8.2=pypi_0
python_abi=3.9=2_cp39
pytz=2023.3=pypi_0
pyzmq=25.1.2=py39h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scanpy=1.9.3=pypi_0
scikit-learn=1.3.0=pypi_0
scikit-learn-extra=0.3.0=pypi_0
scipy=1.11.1=pypi_0
seaborn=0.12.2=pypi_0
session-info=1.0.0=pypi_0
setuptools=75.1.0=py39h06a4308_0
shap=0.41.0=pypi_0
six=1.16.0=pypi_0
slicer=0.0.7=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
statsmodels=0.14.0=pypi_0
stdlib-list=0.9.0=pypi_0
sympy=1.12=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.10.0=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.27.0=pypi_0
termcolor=2.0.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.14=h39e8969_0
tomli=1.2.1=pypi_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.1=py39hb9d737c_3
tqdm=4.65.0=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
triton=2.0.0=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023.3=pypi_0
umap-learn=0.5.3=pypi_0
unioncom=0.4.0=pypi_0
urllib3=2.0.4=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
werkzeug=2.2.2=pypi_0
wheel=0.41.1=pypi_0
wrapt=1.14.1=pypi_0
xz=5.4.6=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_1
